+++ documentation/docs/changelog.md
@@ -1,4 +1,16 @@
+---
+sidebar_position: 2
+---
+
 # Changelog
+
+## 2024.07.19 - version 1.9.0
+
+* A complete rewrite of the Window class content management to introduce 
+  [content pages](resources/views/window#adding-content-to-the-window)
+* Add a count utility method to the Arr class
+* Add the [Callback Loader](resources/core/callback-loader) mechanism to
+  invoke functions after the library is fully loaded
 
 ## 2024.07.10 - version 1.8.0